pyproject.toml
[project]
name = "food-image-recognition"
version = "0.1.0"
description = ""
authors = [
    {name = "cuderbk",email = "richardbui26k4@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10.12"
dependencies = [
    "tensorflow (>=2.19.0,<3.0.0)",
    "kafka-python (>=2.1.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "minio (>=7.2.15,<8.0.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
